requirements.txt
sentencepiece
transformers
torch
fastapi
starlette
aiofiles
python-multipart
Pillow
uvicorn
streamlit
requestspip 